Convert archive.org URLs for <img> tags to https dynamically at runtime.
Some URLs are stored in the database as http. But when they are used
for things like <img> this can cause the web browser to give scary
warnings about mixed encrypted/un-encrypted content.

This fix has been applied to the cover art thumbnail in the browse title
blade, cover art on the main catalog page for a project and the
/api/lastest_releases results used by the Wordpress latest releases
feed.

Ideally the items in the database should probably be updated and the
cataloging process updated to force https for future projects. But that
would require a lot more effort so this workaround is better than nothing
for now.